+++ lottery-data/src/test/java/de/alpharogroup/lottery/box/LotteryBoxTest.java
@@ -21,14 +21,12 @@
 package de.alpharogroup.lottery.box;
 
 import static org.testng.Assert.assertNotNull;
-import static org.testng.AssertJUnit.assertEquals;
 
-import de.alpharogroup.evaluate.object.evaluators.EqualsHashCodeAndToStringEvaluator;
-import lombok.SneakyThrows;
 import org.meanbean.test.BeanTester;
 import org.testng.annotations.Test;
 
 import de.alpharogroup.collections.set.SetFactory;
+import de.alpharogroup.evaluate.object.verifier.ContractVerifier;
 import de.alpharogroup.lottery.enums.LotteryGameType;
 
 /**
@@ -42,15 +40,9 @@ public class LotteryBoxTest
 	 * {@link LotteryBox#toString()}
 	 */
 	@Test
-	@SneakyThrows
-	public void testEqualsHashcodeAndToStringWithClassSilently()
+	public void verifyEqualsHashcodeAndToStringContracts()
 	{
-		boolean expected;
-		boolean actual;
-		actual = EqualsHashCodeAndToStringEvaluator
-			.evaluateEqualsHashcodeAndToString(LotteryBox.class);
-		expected = true;
-		assertEquals(expected, actual);
+		ContractVerifier.of(LotteryBox.class).verify();
 	}
 
 	/**