+++ lottery-data/src/test/java/de/alpharogroup/lottery/drawings/DrawnEuroMillionsNumbersTest.java
@@ -24,16 +24,14 @@
 package de.alpharogroup.lottery.drawings;
 
 import static org.testng.Assert.assertNotNull;
-import static org.testng.AssertJUnit.assertEquals;
 
 import java.util.Set;
 
-import de.alpharogroup.evaluate.object.evaluators.EqualsHashCodeAndToStringEvaluator;
-import lombok.SneakyThrows;
 import org.meanbean.test.BeanTester;
 import org.testng.annotations.Test;
 
 import de.alpharogroup.collections.set.SetFactory;
+import de.alpharogroup.evaluate.object.verifier.ContractVerifier;
 
 /**
  * The unit test class for the class {@link DrawnEuroMillionsNumbers}.
@@ -46,15 +44,9 @@ public class DrawnEuroMillionsNumbersTest
 	 * {@link DrawnEuroMillionsNumbers#hashCode()} and {@link DrawnEuroMillionsNumbers#toString()}
 	 */
 	@Test
-	@SneakyThrows
-	public void testEqualsHashcodeAndToStringWithClassSilently()
+	public void verifyEqualsHashcodeAndToStringContracts()
 	{
-		boolean expected;
-		boolean actual;
-		actual = EqualsHashCodeAndToStringEvaluator
-				.evaluateEqualsHashcodeAndToString(DrawnEuroMillionsNumbers.class);
-		expected = true;
-		assertEquals(expected, actual);
+		ContractVerifier.of(DrawnEuroMillionsNumbers.class).verify();
 	}
 
 	/**