+++ lottery-data/src/test/java/de/alpharogroup/lottery/drawings/DrawnLotteryNumbersTest.java
@@ -21,16 +21,14 @@
 package de.alpharogroup.lottery.drawings;
 
 import static org.testng.Assert.assertNotNull;
-import static org.testng.AssertJUnit.assertEquals;
 
 import java.util.Set;
 
-import de.alpharogroup.evaluate.object.evaluators.EqualsHashCodeAndToStringEvaluator;
-import lombok.SneakyThrows;
 import org.meanbean.test.BeanTester;
 import org.testng.annotations.Test;
 
 import de.alpharogroup.collections.set.SetFactory;
+import de.alpharogroup.evaluate.object.verifier.ContractVerifier;
 
 /**
  * The unit test class for the class {@link DrawnLotteryNumbers}.
@@ -43,15 +41,9 @@ public class DrawnLotteryNumbersTest
 	 * {@link DrawnLotteryNumbers#hashCode()} and {@link DrawnLotteryNumbers#toString()}
 	 */
 	@Test
-	@SneakyThrows
-	public void testEqualsHashcodeAndToStringWithClassSilently()
+	public void verifyEqualsHashcodeAndToStringContracts()
 	{
-		boolean expected;
-		boolean actual;
-		actual = EqualsHashCodeAndToStringEvaluator
-				.evaluateEqualsHashcodeAndToString(DrawnLotteryNumbers.class);
-		expected = true;
-		assertEquals(expected, actual);
+		ContractVerifier.of(DrawnLotteryNumbers.class).verify();
 	}
 
 	/**