+++ lottery-data/src/test/java/de/alpharogroup/lottery/played/PlayedLotteryTicketsTest.java
@@ -20,7 +20,6 @@
  */
 package de.alpharogroup.lottery.played;
 
-import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertNotNull;
 
 import java.util.Set;
@@ -29,9 +28,8 @@ import org.meanbean.test.BeanTester;
 import org.testng.annotations.Test;
 
 import de.alpharogroup.collections.set.SetFactory;
-import de.alpharogroup.evaluate.object.evaluators.EqualsHashCodeAndToStringEvaluator;
+import de.alpharogroup.evaluate.object.verifier.ContractVerifier;
 import de.alpharogroup.lottery.ticket.LotteryTicket;
-import lombok.SneakyThrows;
 
 /**
  * The unit test class for the class {@link PlayedLotteryTickets}.
@@ -43,16 +41,10 @@ public class PlayedLotteryTicketsTest
 	 * Test method for {@link PlayedLotteryTickets#equals(Object)} ,
 	 * {@link PlayedLotteryTickets#hashCode()} and {@link PlayedLotteryTickets#toString()}
 	 */
-	@SneakyThrows
 	@Test
-	public void testEqualsHashcodeAndToStringWithClass()
+	public void verifyEqualsHashcodeAndToStringContracts()
 	{
-		boolean expected;
-		boolean actual;
-		actual = EqualsHashCodeAndToStringEvaluator
-			.evaluateEqualsHashcodeAndToString(PlayedLotteryTickets.class);
-		expected = true;
-		assertEquals(expected, actual);
+		ContractVerifier.of(PlayedLotteryTickets.class).verify();
 	}
 
 	/**