+++ CHANGELOG.md
@@ -6,11 +6,13 @@ Version 5.5-SNAPSHOT
 
 ADDED:
  
-- new dependency of jobj-contract-verifier in version 3 added
+- new dependency of jobj-contract-verifier in version 3.2 added
 
 CHANGED:
 
-- update of parent version to 4.8
+- update of parent version to 5
+- update of dependency silly-collections version to 5.2.1
+- update of randomizer version to 6.0.1
 
 Version 5.4
 -------------